+++ frontend/cypress/e2e/database.cy.ts
@@ -1,5 +1,5 @@
+import { User } from "../../src/types/interfaces";
 import { Country } from "../../src/types/interfaces";
-
 describe("Users API Tests", () => {
   it("fetches all users from the database", () => {
     cy.request("GET", "/api/users").then((response) => {
@@ -7,32 +7,19 @@ describe("Users API Tests", () => {
       expect(response.body).to.be.an("array");
     });
   });
-
   it("adds a new user to the database", () => {
-    const uniqueEmail = `testuser${Date.now()}@example.com`;
-    cy.request({
-      method: "POST",
-      url: "http://localhost/api/users",
-      body: {
-        first_name: "Test",
-        last_name: "User",
-        email: uniqueEmail,
-        password: "password123",
-      },
-      failOnStatusCode: false,
-    }).then((response) => {
-      if (
-        response.status === 400 &&
-        response.body.error === "Email already exists"
-      ) {
-        cy.log("Email already exists");
-      } else {
-        expect(response.status).to.eq(201);
-      }
+    const newUser: User = {
+      first_name: "Test",
+      last_name: "User",
+      email: "newuser@example.com",
+      password: "password123",
+    };
+    cy.request("POST", "/api/users", newUser).then((response) => {
+      expect(response.status).to.eq(201); // Created
+      expect(response.body).to.have.property("first_name", newUser.first_name);
     });
   });
 });
-
 describe("Countries API Tests", () => {
   it("fetches all countries from the database", () => {
     cy.request("GET", "/api/countries").then((response) => {
@@ -41,9 +28,8 @@ describe("Countries API Tests", () => {
     });
   });
   it("adds a new country to the database", () => {
-    const uniqueCountryName = `Testland${Date.now()}`;
     const newCountry: Country = {
-      country_name: uniqueCountryName,
+      country_name: "Testland",
       country_description: "A fictional country for testing.",
       country_capital: "Test City",
       country_population: 12345,
@@ -52,7 +38,6 @@ describe("Countries API Tests", () => {
       country_currency: "Test Currency",
       country_flag: "http://example.com/flag.png",
     };
-
     cy.request("POST", "/api/countries", newCountry).then((response) => {
       expect(response.status).to.eq(201);
       expect(response.body).to.have.property(
@@ -61,36 +46,14 @@ describe("Countries API Tests", () => {
       );
     });
   });
-
   it("fetches details of a specific country", () => {
-    const uniqueCountryName = `Testland${Date.now()}`;
-    const newCountry: Country = {
-      country_name: uniqueCountryName,
-      country_description: "A fictional country for testing.",
-      country_capital: "Test City",
-      country_population: 12345,
-      country_continent: "Test Continent",
-      country_language: "Test Language",
-      country_currency: "Test Currency",
-      country_flag: "http://example.com/flag.png",
-    };
-
-    cy.request("POST", "/api/countries", newCountry).then(() => {
-      cy.request("GET", `/api/countries/${uniqueCountryName}`).then(
-        (response) => {
-          expect(response.status).to.eq(200);
-          expect(response.body).to.have.property(
-            "country_name",
-            uniqueCountryName
-          );
-        }
-      );
+    cy.request("GET", "/api/countries/Testland").then((response) => {
+      expect(response.status).to.eq(200);
+      expect(response.body).to.have.property("country_name", "Testland");
     });
   });
 });
-
 describe("Travel List API Tests", () => {}); //TODO
-
 describe("Visit Status API Tests", () => {
   it("fetches all visit statuses from the database", () => {
     cy.request("GET", "/api/visitstatus").then((response) => {