+++ frontend/src/components/Footer.tsx
@@ -1,0 +1,9 @@
+export default function Footer() {
+  return (
+    <footer className='bg-dark text-white text-center py-3'>
+      <div>
+        <p className='mb-0'>World Compass &copy; 2024</p>
+      </div>
+    </footer>
+  );
+}