+++ frontend/src/components/NavBar.tsx
@@ -1,0 +1,28 @@
+import { Link } from "react-router-dom";
+import { Navbar as BootstrapNavbar, Nav, Container } from "react-bootstrap";
+
+export default function Navbar() {
+  return (
+    <BootstrapNavbar bg='dark' variant='dark' expand='lg'>
+      <Container>
+        <BootstrapNavbar.Brand as={Link} to='/'>
+          World Compass
+        </BootstrapNavbar.Brand>
+        <BootstrapNavbar.Toggle aria-controls='basic-navbar-nav' />
+        <BootstrapNavbar.Collapse id='basic-navbar-nav'>
+          <Nav className='ms-auto'>
+            <Nav.Link as={Link} to='/'>
+              Home
+            </Nav.Link>
+            <Nav.Link as={Link} to='/loginreg'>
+              Login/Register
+            </Nav.Link>
+            <Nav.Link as={Link} to='/loginreg'>
+              Log out
+            </Nav.Link>
+          </Nav>
+        </BootstrapNavbar.Collapse>
+      </Container>
+    </BootstrapNavbar>
+  );
+}