+++ frontend/src/pages/HomePage.tsx
@@ -1,0 +1,10 @@
+import Countries from "../components/Countries";
+
+export default function HomePage() {
+  return (
+    <>
+      <h1>Countries</h1>
+      <Countries />{" "}
+    </>
+  );
+}