+++ frontend/src/pages/LoginRegPage.tsx
@@ -1,0 +1,7 @@
+export default function LoginRegPage() {
+  return (
+    <>
+      <h1>Login Register</h1>
+    </>
+  );
+}